Split /work into /writing and /building
Natalie found the stacked Work page buried the projects below the
research list. The two content types are now peers in the nav: /writing
holds the research index, /building holds the inkling-style project
blocks, each with its own hero and everything above the fold. All legacy
routes (/work, /research, /projects, /products) redirect to the right
new page; nav, palette, homepage copy, and llms.txt updated.
