+++ app/products/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation"
 
 export default function ProductsRedirect() {
-  redirect("/work")
+  redirect("/building")
 }
