+++ app/projects/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation"
 
 export default function ProjectsRedirect() {
-  redirect("/work")
+  redirect("/building")
 }
