+++ app/research/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation"
 
 export default function ResearchRedirect() {
-  redirect("/work")
+  redirect("/writing")
 }
